方块查找跳过全空气子区块

+++ src/main/java/boat/carpetorgaddition/wheel/ProgressBar.java
@@ -16,7 +16,7 @@ public class ProgressBar {
         this.end = end;
     }
 
-    public void setProgress(int current) {
+    public void setProgress(long current) {
         this.progress = Math.min(MathUtils.normalize(current, 0.0, this.end), 1);
     }
